==> 1 preprocessing_AMI.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

'''
input = raw data
output = raw_train, raw_test
function:The raw data was devided into raw_train and raw_test
'''

df = pd.read_excel('raw data.xlsx')
# removing patients without follow-up result
df = df[df['MACE'].notnull()]

df = df.reindex(np.random.permutation(df.index))#disrupt the order

df.shape

df.head()

df['MACE'].value_counts()

df.columns

a = df[df['admission number']==1025711].index.tolist()[0]
b = df[df['admission number']==1057687].index.tolist()[0]
#Two patients (admission number:1025711,1057687) were excluded because they didn't make multiple tests
df = df.drop(index=[a,b])

df.shape

raw_train = df[0:244]

raw_train['MACE'].value_counts()

raw_test = df[244:408]
raw_test['MACE'].value_counts()

#output:raw_train, raw_test
raw_train.to_excel(r'C:\Users\lenovo\Revision\Preprocessing\raw_train.xlsx')
raw_test.to_excel(r'C:\Users\lenovo\Revision\Preprocessing\raw_test.xlsx')
