==> 2 raw_train preprocessing.ipynb <==
#input:raw_train.xlsx
#output:processed_train.xlsx
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

raw_train = pd.read_excel('raw_train.xlsx',index_col=0)

raw_train.head()

#Encode gender(man equal to 0,women equal to 0)
raw_train.loc[raw_train['Sex'] == '男', 'Sex'] = 0
raw_train.loc[raw_train['Sex'] == '女', 'Sex'] = 1

#Calculate the number of days after discharge(follow-up time = follow-up data - hospital date)
import datetime
import time
follow_date = datetime.datetime(2020,11,1,0,0,0)
ts = pd.DatetimeIndex([follow_date])
for i in raw_train.index:
    a = ts - raw_train['InhisTime'][i]
    b = a.days
    raw_train['InhisTime'][i] = b[0]        
raw_train.rename(columns={'InhisTime':'Follow_up time'},inplace=True)

raw_train.head()

raw_train.isnull().sum()

'''
input:a discrete variable
output:one of all discrete variable categories
function:generating one of all discrete variable categories according to the proportion of this categort
For example,the number of man(labelled as 0) is 144,the number of woman(labelled as 0) is 100,
then this function has a probability of 144/(144+100) for producing the value 0
'''
def randnum(data):
    a = np.random.random()
    count_values = (data.value_counts()).values
    count_index = (data.value_counts()).index
    notnull_datalen = data.notnull().sum()
    b = []
    sum_value = 0
    return_value = []
    for i in range(len(data.value_counts())):
        sum_value += count_values[i]
        b.append(sum_value/notnull_datalen)
        if a < b[i]:
            return_value = count_index[i]
            break
    return return_value

#Discrete variable imputation
#OtherHis imputation
otherhis = raw_train['OtherHis']
otherhis_is_null = pd.isnull(otherhis)
otherhis_null_true = otherhis[otherhis_is_null]
raw_train['OtherHis'][otherhis_null_true.index] = randnum(raw_train['OtherHis'])

#Drugcom imputation
Drugcom = raw_train['Drugcom']
Drugcom_is_null = pd.isnull(Drugcom)
Drugcom_null_true = Drugcom[Drugcom_is_null]
for index_null in Drugcom_null_true.index:
    raw_train['Drugcom'][index_null] = randnum(raw_train['Drugcom'])

#Culp imputation
Culp = raw_train['Culp']
Culp_is_null = pd.isnull(Culp)
Culp_null_true = Culp[Culp_is_null]
for index_null in Culp_null_true.index:
    raw_train['Culp'][index_null] = randnum(raw_train['Culp'])

#MainVASc imputation
MainVASc = raw_train['MainVASc']
MainVASc_is_null = pd.isnull(MainVASc)
MainVASc_null_true = MainVASc[MainVASc_is_null]
for index_null in MainVASc_null_true.index:
    raw_train['MainVASc'][index_null] = randnum(raw_train['MainVASc'])

#Numsten imputation
Numsten = raw_train['Numsten']
Numsten_is_null = pd.isnull(Numsten)
Numsten_null_true = Numsten[Numsten_is_null]
raw_train['Numsten'][Numsten_null_true.index] = randnum(raw_train['Numsten'])

#The imputation of discrete variables is complete
raw_train.isnull().sum()

'''
input:a continuous variable
output:None
function:imputing a continuous variable with the values generated by Gaussian distribution.
if the generated values are not in the range of the continuous variable,the variable will be imputed with the mean of this variable
'''
def fill_null(data):
    data_is_null = pd.isnull(data)
    data_null_true = data[data_is_null]
    for index_null in data_null_true.index:
        gauss = np.random.normal(np.mean(data),np.var(data),1)
        if gauss > raw_train[data.name].max():
            raw_train[data.name][index_null] = np.mean(data)
        elif gauss < raw_train[data.name].min():
            raw_train[data.name][index_null] = np.mean(data)
        else:
            raw_train[data.name][index_null] = gauss

postive_train = raw_train[raw_train['MACE'] == 1][:]
negative_train = raw_train[raw_train['MACE'] == 0][:]

#imputation of TC
postive_TC = postive_train['TC']
negative_TC = negative_train['TC']
fill_null(postive_TC)
fill_null(negative_TC)

#imputation of LDL
postive_LDL = postive_train['LDL']
negative_LDL = negative_train['LDL']
fill_null(postive_LDL)
fill_null(negative_LDL)

#imputation of CRP
postive_CRP = postive_train['CRP']
negative_CRP = negative_train['CRP']
fill_null(postive_CRP)
fill_null(negative_CRP)

#imputation of EF
postive_EF = postive_train['EF']
negative_EF = negative_train['EF']
fill_null(postive_EF)
fill_null(negative_EF)

#imputation of LV
postive_LV = postive_train['LV']
negative_LV = negative_train['LV']
fill_null(postive_LV)
fill_null(negative_LV)

#imputation of Cr
postive_Cr = postive_train['Cr']
negative_Cr = negative_train['Cr']
fill_null(postive_Cr)
fill_null(negative_Cr)

#imputation of UA
postive_UA = postive_train['UA']
negative_UA = negative_train['UA']
fill_null(postive_UA)
fill_null(negative_UA)

#imputation of Glu
postive_Glu = postive_train['Glu']
negative_Glu = negative_train['Glu']
fill_null(postive_Glu)
fill_null(negative_Glu)

#imputation of WBC
postive_WBC = postive_train['WBC']
negative_WBC = negative_train['WBC']
fill_null(postive_WBC)
fill_null(negative_WBC)

#imputation of N
postive_N = postive_train['N']
negative_N = negative_train['N']
fill_null(postive_N)
fill_null(negative_N)

#The imputation of continuous and discrete variables is complete
raw_train.isnull().sum()

#input：strings of 'Revasc(Time of revascularization)'
#output：minutes
#d,h,m convert to minutes
def time_conv(strings):
    day = 0
    hour = 0
    mins = 0
    if 'd' in strings:
        d_end = strings.index('d')
        day = int(strings[0:d_end])
    if 'h' in strings:
        if 'd' in strings:
            h_star = strings.index('d')+1
            h_end = strings.index('h')
            hour = int(strings[h_star:h_end])
        else:
            h_end = strings.index('h')
            hour = int(strings[0:h_end])
    if 'm' in strings:
        if 'h' in strings:
            m_star = strings.index('h')+1
            m_end = strings.index('m')
            mins = int(strings[m_star:m_end])
        elif 'd' in strings:
            m_star = strings.index('d')+1
            m_end = strings.index('m')
            mins = int(strings[m_star:m_end])
        else:
            m_end = strings.index('m')
            mins = int(strings[0:m_end])
    if day > 3:
        day = 3
    time_mins = day*1440+hour*60+mins
    return time_mins

for i in raw_train.index:
    conv = raw_train['Revasc'][i] 
    raw_train['Revasc'][i] = time_conv(conv)
raw_train['Revasc'] = raw_train['Revasc'].astype(int)

raw_train.head()

for i in raw_train.index:
    if raw_train['Killip'][i] == 'I':
        raw_train['Killip'][i] = 1
    elif raw_train['Killip'][i] == 'Ⅰ':
        raw_train['Killip'][i] = 1
    elif raw_train['Killip'][i] == 'I ':
        raw_train['Killip'][i] = 1
    elif raw_train['Killip'][i] == 'II':
        raw_train['Killip'][i] = 2
    elif raw_train['Killip'][i] == 'Ⅱ':
        raw_train['Killip'][i] = 2
    elif raw_train['Killip'][i] == 'III':
        raw_train['Killip'][i] = 3
    elif raw_train['Killip'][i] == 'Ⅲ':
        raw_train['Killip'][i] = 3
    elif raw_train['Killip'][i] == 'IV':
        raw_train['Killip'][i] = 4
    elif raw_train['Killip'][i] == 'Ⅳ':
        raw_train['Killip'][i] = 4

raw_train['Killip'] = raw_train['Killip'].astype(int)

raw_train['Sex'] = raw_train['Sex'].astype(int)
raw_train['Follow_up time'] = raw_train['Follow_up time'].astype(int)

raw_train = raw_train.astype(float)

raw_train.to_excel(r'C:\Users\lenovo\Revision\Preprocessing\processed_train.xlsx')

raw_train.dtypes
